% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp11examplesep26-package.R
\docType{package}
\name{cpp11examplesep26-package}
\alias{cpp11examplesep26}
\alias{cpp11examplesep26-package}
\title{cpp11examplesep26: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
